Fix queue for api and add queue system
